+++ tools/test-groups.mjs
@@ -23,7 +23,11 @@ const GROUPS = Object.freeze({
       /corrective-1310\.test\.mjs$/i,
       /corrective-1313\.test\.mjs$/i,
       /optimization-13304\.test\.mjs$/i,
-      /optimization-13306\.test\.mjs$/i
+      /optimization-13306\.test\.mjs$/i,
+      /optimization-13307\.test\.mjs$/i,
+      /optimization-13308\.test\.mjs$/i,
+      /optimization-13309\.test\.mjs$/i,
+      /optimization-133010\.test\.mjs$/i
     ]
   },
   newtab: {
@@ -67,7 +71,10 @@ const GROUPS = Object.freeze({
       /corrective-13218\.test\.mjs$/i,
       /optimization-13303\.test\.mjs$/i,
       /optimization-13304\.test\.mjs$/i,
-      /optimization-13306\.test\.mjs$/i
+      /optimization-13306\.test\.mjs$/i,
+      /optimization-13307\.test\.mjs$/i,
+      /optimization-13308\.test\.mjs$/i,
+      /optimization-13309\.test\.mjs$/i
     ]
   },
   sync: {
@@ -96,7 +103,8 @@ const GROUPS = Object.freeze({
       /corrective-13215\.test\.mjs$/i,
       /corrective-13216\.test\.mjs$/i,
       /corrective-13217\.test\.mjs$/i,
-      /optimization-13303\.test\.mjs$/i
+      /optimization-13303\.test\.mjs$/i,
+      /optimization-133010\.test\.mjs$/i
     ]
   },
   recovery: {
@@ -201,7 +209,8 @@ const GROUPS = Object.freeze({
       /corrective-13215\.test\.mjs$/i,
       /corrective-13216\.test\.mjs$/i,
       /corrective-13218\.test\.mjs$/i,
-      /process-13305\.test\.mjs$/i
+      /process-13305\.test\.mjs$/i,
+      /optimization-133010\.test\.mjs$/i
     ]
   }
 });